requirements.txt
alembic==1.13.1
aniso8601==9.0.1
Flask==3.0.3
Flask-Cors==4.0.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
SQLAlchemy==2.0.30
mysql-connector-python==9.0.0
requests==2.32.3